replace source region floats with doubles to handle small fluxes appropriately

+++ include/openmc/random_ray/random_ray.h
@@ -70,17 +70,17 @@ public:
 
   //----------------------------------------------------------------------------
   // Public data members
-  vector<float> angular_flux_;
+  vector<double> angular_flux_;
   bool ray_trace_only_ {false}; // If true, only perform geometry operations
 
   //---------------------------------------------------------------------------
   // Public data members for kinetic simulations
-  vector<float> angular_flux_prime_;
+  vector<double> angular_flux_prime_;
 
 private:
   //----------------------------------------------------------------------------
   // Private data members
-  vector<float> delta_psi_;
+  vector<double> delta_psi_;
   vector<MomentArray> delta_moments_;
   vector<int> mesh_bins_;
   vector<double> mesh_fractional_lengths_;
